# desktop/src/lib/openapi.json
{
  "components": {
    "schemas": {
...
      "ConfigResponse": {
        "properties": {
          "anthropic": {
            "$ref": "#/components/schemas/APIKeyMaskedSettings"
          },
          "argos": {
            "$ref": "#/components/schemas/APIKeyMaskedSettings"
          },
          "debug_mode": {
            "title": "Debug Mode",
            "type": "boolean"
          },
          "gemini": {
            "$ref": "#/components/schemas/APIKeyMaskedSettings"
          },
          "gui": {
            "$ref": "#/components/schemas/GUISettings"
          },
          "openai": {
            "$ref": "#/components/schemas/APIKeyMaskedSettings"
          },
          "processing": {
            "$ref": "#/components/schemas/ProcessingSettings"
          },
          "rag": {
            "$ref": "#/components/schemas/RAGSettings"
          },
          "translation": {
            "$ref": "#/components/schemas/TranslationConfig"
          }
        },
        "required": [
          "openai",
          "gemini",
          "anthropic",
          "argos",
          "translation",
          "rag",
          "gui",
          "processing",
          "debug_mode"
        ],
        "title": "ConfigResponse",
        "type": "object"
      },
...
      "EndpointModelsResponse": {
        "description": "The models a service's saved endpoint says it serves.\n\nA failure is `error`, not an HTTP error: a local server that isn't running\nyet is ordinary, and the picker still offers the saved model without it.",
        "properties": {
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "models": {
            "items": {
              "type": "string"
            },
            "title": "Models",
            "type": "array"
          }
        },
        "title": "EndpointModelsResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "PDFusion sidecar",
    "version": "1.2.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/config": {
      "get": {
        "operationId": "get_config_config_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Config",
        "tags": [
          "config"
        ]
      },
      "put": {
        "operationId": "update_config_config_put",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Config",
        "tags": [
          "config"
        ]
      }
    },
...
    "/config/models/{service}": {
      "get": {
        "description": "The ids the saved key can use at the saved endpoint.\n\nA wrapper over the catalog (`GET /providers/{id}/models`) for the\ntoolbar's model picker: listed models only \u2014 the picker adds the saved\nmodel and, without a list, the suggestions itself. Always the saved key\nwith the saved endpoint, the pair `PUT /config` keeps together, so this\ncan't send a key anywhere new.",
        "operationId": "list_endpoint_models_config_models__service__get",
        "parameters": [
          {
            "in": "path",
            "name": "service",
            "required": true,
            "schema": {
              "description": "A provider's id, as GET /providers lists them",
              "title": "Service",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EndpointModelsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Endpoint Models",
        "tags": [
          "config"
        ]
      }
    },
...
    "/config/validate": {
      "post": {
        "description": "Check credentials with the provider, by listing the key's models.\n\nA wrapper over `POST /providers/{id}/verify`. Whatever the request leaves\nout comes from the saved settings: the key, the model, the endpoint. The\nsaved key is only sent to the saved endpoint, the rule `PUT /config`\nkeeps; checking another endpoint needs the key typed alongside it.",
        "operationId": "validate_credentials_config_validate_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Validate Credentials",
        "tags": [
          "config"
        ]
      }
    },
...
    "/providers": {
      "get": {
        "description": "Every provider, with where its key stands. Reads only; never lists.",
        "operationId": "list_providers_providers_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProvidersResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Providers",
        "tags": [
          "providers"
        ]
      }
    },
    "/providers/{provider_id}": {
      "put": {
        "description": "Save one provider's key, endpoint, models and parameters.\n\nSaves without checking the key with the provider: `POST .../verify` is\nthat step, and \"Save anyway\" has to exist. Unlike `PUT /config`, a first\nkey saved here does not move translation off Argos by itself; the caller\nchooses with `PUT /config`'s `translation_model`.",
        "operationId": "update_provider_providers__provider_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "provider_id",
            "required": true,
            "schema": {
              "description": "A provider's id, as GET /providers lists them",
              "title": "Provider Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Provider",
        "tags": [
          "providers"
        ]
      }
    },
    "/providers/{provider_id}/key": {
      "delete": {
        "description": "Forget the saved key, a preserved unreadable one included. The\nendpoint stays; a key entered later goes to it.",
        "operationId": "delete_provider_key_providers__provider_id__key_delete",
        "parameters": [
          {
            "in": "path",
            "name": "provider_id",
            "required": true,
            "schema": {
              "description": "A provider's id, as GET /providers lists them",
              "title": "Provider Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Provider Key",
        "tags": [
          "providers"
        ]
      }
    },
    "/providers/{provider_id}/models": {
      "get": {
        "operationId": "provider_models_providers__provider_id__models_get",
        "parameters": [
          {
            "in": "path",
            "name": "provider_id",
            "required": true,
            "schema": {
              "description": "A provider's id, as GET /providers lists them",
              "title": "Provider Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "refresh",
            "required": false,
            "schema": {
              "default": false,
              "title": "Refresh",
              "type": "boolean"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelCatalogResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Provider Models",
        "tags": [
          "providers"
        ]
      }
    },
...
  }
}